Update expense table seeder
This commit adds more depth to the expense table seeder.  This is
accomplished by using faker, which allows for the generation of
arbitrary numbers of records as opposed to hard coded records.  This
does come at the cost of less realistic data.  Notably, this seeder
isn't able to provide sensible fixed monthly expenses such as mortgage,
taxes, and car payments as well as not being able to guarantee that
expenses are present in every month.  Descriptions of expenses are also
not sensible.  At this point, the simplicity of the seeder outweighs
the potentially less realistic data.