Add per-month analytics for expenses
This commit adds in functionality to display analytics for expenses
by month and year.  This was accomplished mainly by adding in local
query scopes in the Expense class.  The corresponding controller and
view have been updated to reflect the change to monthly analytics.
Routes have also been added so that the URL expresses the month and
year.