Update per-month analytics for expenses
This commit stops using the local query scopes defined in the Expense
model in favor of parsing the data in the controller.  The chart data
is also built differently.  Namely, an object with all expense types
is initialized and monthly data is added to it in a loop.  This is in
contrast to starting with expense types with expenses attached,
filtering the expenses by month, and executing the aggregate sum
functions to give totals.  This should be a more semantically natural
way to parse the data.