Create add expense pages
The page for adding expenses has been created as well as the storage
controller and corresponding routing.  Error messages are presented
to the user upon failure to pass validation, and entries from the
invalid attempt are used to repopulate the form.  The month and year
combination is preserved throughout the process.  That is, when a
user clicks the add expense button for a particular month, the add
expense form pre-populates the date field with the month and year.
Upon successful completion of the expense add form, the user is
redirected to the overview page for the date specified in the
expense.