Add expense edit
The ability to edit expenses has been added.  A user can click on an
expense in the table, and they are taken to the expense edit page.
The edit page fully validates and saves the expense upon successful
validation.  The user is returned to the overview page for the date
provided in the edit.