Update form_functions.js
The form_functions.js file has been updated to target classes instead
ids.  This will allow multiple areas to be updated more naturally.  It
also removes .display-none when the trigger has a value of 1 and adds
the .display-none when the trigger has a value of 0.  This is in
contrast to toggling the .display-none class.