Update overview
The historical spending graph has been removed.  In it's place there
is a table that lists the expense types that have a monthly budget
target.  The table lists the spending to date, monthly budget target,
and available spending (monthly budget target minus the spending to
date).  The position of the historical spending chart (now expense
table) and the current month spending chart are swapped.  This allows
the legend to stick with the spending chart for smaller breakpoints.