Fix clickable table
The clickable table functionality has been updated.  The trigger
initially fired on tr, and has been updated to fire on .clickable td.
The navigation to find the correct id has also been updated.  This
fixes the issue of being able to click on table headings of tables
with the clickable class, and clicking on tables without the
clickable class.