Add available funds to overview chart
The available funds have been incorporated into the overview pie
chart.  This includes calculating available funds in the controller.
It is important to note that the value will go as low as zero; it
will not be negative.  The chart seems to display the absolute value
of an entry regardless of the sign.