Add at_most information to overview
The at_most property has been incorporated into the overview page.
This includes indicating at most or at least for the monthly budget
column as well as an over budget notice on the available spending.
The over budget notice allows for all values to be displayed as
non-negative allowing for the dollar sign to be used.  With the
negative before, formatting was problematic.  Table rows are also
color coded based on the spending to date and monthly budget in
conjunction with the at_most property.