Update Income migration and seeder
The Income model has been updated to reflect the newly introduced
IncomeType model.  This includes an updated migration and seeder.  In
addition to adding IncomeType data to the Income seeder, the Income
seeder has been updated to give a more realistic monthly income as
opposed to randomly assigning income values across the year.