Implement CRUD for IncomeTypes
The dummy endpoints for the IncomeTypes model have been implemented
and are entirely functional.  All endpoints reside in the
budget_settings group.  The display table for the IncomeTypes has
also been updated to properly present the Monthly Income Amount and
Recurring End Date.  The behavior before would show extra and
incorrect data for the Monthly Income Amount as well as the Recurring
End Date.